basic app restructuring
