+++ requirements.txt
@@ -13,3 +13,5 @@ streamlit-drawable-canvas==0.9.2
 numpy==1.24.1
 mesa
 streamlit-javascript==0.1.5
+psycopg2==2.9.5
+psycopg2-binary==2.9.5
